+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Contao core bundle change log
 
+### 4.3.3 (2016-01-23)
+
+ * Fix removing the pagination limit in the back end (see #675).
+ * Fix the version contraints of the manager plugin.
+
 ### 4.3.2 (2017-01-18)
 
  * Correctly handle nested public folders when symlinking a folder.